Add files via upload
In my first hackathon, we were given a song to inspire our project. The song that my team received was from the movie Coco, 'Remember Me.' So, we created a web app that predicts the time you have left to live and where you will go when you die. It was just for fun and does not represent real life.